+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##  Unreleased
 
+* Add `From` trait to coerce `candid::Error` into `ic_agent::AgentError`.
+* Add `Agent::set_arc_identity` method to switch identity.
+
 ## [0.26.1] - 2023-08-22
 
 Switched from rustls crate to rustls-webpki fork to address https://rustsec.org/advisories/RUSTSEC-2023-0052